ai/requirements.txt
# KU Prep Arena — AI / Dataset Section
# Run: pip install -r requirements.txt

# Core
openai>=1.35.0          # OpenAI-compatible client for vLLM endpoint
httpx>=0.27.0

# PDF & text
pdfplumber>=0.11.0
pymupdf>=1.24.0         # fitz — fast PDF text extraction

# Dataset & eval
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0     # TF-IDF cosine similarity

# Visualisation (notebooks)
matplotlib>=3.9.0
seaborn>=0.13.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
